pyproject.toml
[project]
name = "appscrip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.2",
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "uvicorn>=0.41.0",
]
